feat: 알람 설정 시 모달 적용 / 관련 코드 model 폴더로 분리
