+++ src/widgets/mypage/notification-settings/model/constant/notification-items.ts
@@ -1,0 +1,12 @@
+export const NOTIFICATION_ITEMS = [
+  {
+    id: 'service',
+    title: '서비스 알림',
+    description: '팀장 선정 결과, 팀 규칙 변경 등 필수 정보를 알려드려요.',
+  },
+  {
+    id: 'sms',
+    title: '문자(SMS) 알림',
+    description: '앱을 확인하지 않을 때도 중요한 소식을 문자로 보내드려요.',
+  },
+] as const;
